@@ -5,14 +5,14 @@ plugins {
 
 android {
     namespace 'com.liamgooch.awis_phonebook'
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.liamgoooch.awis_phonebook"
         minSdk 24
-        targetSdk 33
-        versionCode 1
-        versionName "1.0"
+        targetSdk 34
+        versionCode 3
+        versionName "1.1"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
