@@ -3,7 +3,7 @@ package com.liamgooch.awis_phonebook
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
 
-import org.junit.Test
+import org.junit.jupiter.api.Test
 import org.junit.runner.RunWith
 
 import org.junit.Assert.*
