@@ -1,6 +1,6 @@
 package com.liamgooch.awis_phonebook
 
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 import org.junit.Assert.*
 
@@ -10,7 +10,7 @@ import org.junit.Assert.*
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 class ExampleUnitTest {
-    @Test
+    @org.junit.jupiter.api.Test
     fun addition_isCorrect() {
         assertEquals(4, 2 + 2)
     }
